requirements.txt
crewai==0.11.2
python-dotenv>=1.0.0
requests>=2.31.0
langchain>=0.0.335
PyPDF2>=3.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
Flask>=2.0.0
Werkzeug>=2.0.0
gunicorn>=20.1.0
urllib3>=1.26.0
tiktoken>=0.5.2,<0.6.0 